hospital/requirements.txt
aiosmtplib==1.1.6
anyio==3.5.0
asgiref==3.5.0
backports.zoneinfo==0.2.1
certifi==2021.10.8
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
Django==4.0.3
django-rest-swagger==2.2.0
djangorestframework==3.13.1
drf-yasg==1.20.0
h11==0.12.0
httpcore==0.13.7
httpx==0.19.0
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.1
MarkupSafe==2.1.1
messages==0.8.0
openapi-codec==1.3.2
packaging==21.3
pyparsing==3.0.7
pytz==2022.1
PyYAML==6.0
requests==2.27.1
rfc3986==1.5.0
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
simplejson==3.17.6
sniffio==1.2.0
sqlparse==0.4.2
tzdata==2022.1
uritemplate==4.1.1
urllib3==1.26.9
validus==0.3.0
